# backend/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserSelfDTO": {
				"properties": {
					"permissions": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"roles": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"isBanned": {
						"type": "boolean"
					},
					"warningCount": {
						"type": "number",
						"format": "double"
					},
					"uploadCount": {
						"type": "number",
						"format": "double"
					},
					"showUploads": {
						"type": "boolean"
					},
					"showFavorites": {
						"type": "boolean"
					},
					"showRatings": {
						"type": "boolean"
					},
					"showComments": {
						"type": "boolean"
					},
					"emailVerifiedAt": {
						"type": "string",
						"nullable": true
					},
					"updatedAt": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"website": {
						"type": "string",
						"nullable": true
					},
					"bio": {
						"type": "string",
						"nullable": true
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"birthDate": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"isBanned",
					"warningCount",
					"uploadCount",
					"showUploads",
					"showFavorites",
					"showRatings",
					"showComments",
					"emailVerifiedAt",
					"updatedAt",
					"createdAt",
					"website",
					"bio",
					"avatarUrl",
					"birthDate",
					"email",
					"username",
					"id"
				],
				"type": "object"
			},
			"UserPublicDTO": {
				"properties": {
					"createdAt": {
						"type": "string"
					},
					"website": {
						"type": "string",
						"nullable": true
					},
					"bio": {
						"type": "string",
						"nullable": true
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"username": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"createdAt",
					"website",
					"bio",
					"avatarUrl",
					"username",
					"id"
				],
				"type": "object"
			},
			"UserAdminDTO": {
				"allOf": [
					{
						"$ref": "#/components/schemas/UserPublicDTO"
					},
					{
						"properties": {
							"permissions": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"roles": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"deletedAt": {
								"type": "string",
								"nullable": true
							},
							"isBanned": {
								"type": "boolean"
							},
							"warningCount": {
								"type": "number",
								"format": "double"
							},
							"uploadCount": {
								"type": "number",
								"format": "double"
							},
							"showUploads": {
								"type": "boolean"
							},
							"showFavorites": {
								"type": "boolean"
							},
							"showRatings": {
								"type": "boolean"
							},
							"showComments": {
								"type": "boolean"
							},
							"emailVerifiedAt": {
								"type": "string",
								"nullable": true
							},
							"updatedAt": {
								"type": "string"
							},
							"birthDate": {
								"type": "string",
								"nullable": true
							},
							"email": {
								"type": "string"
							}
						},
						"required": [
							"permissions",
							"roles",
							"deletedAt",
							"isBanned",
							"warningCount",
							"uploadCount",
							"showUploads",
							"showFavorites",
							"showRatings",
							"showComments",
							"emailVerifiedAt",
							"updatedAt",
							"birthDate",
							"email"
						],
						"type": "object"
					}
				]
			},
			"UserVisibleDTO": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/UserPublicDTO"
					},
					{
						"$ref": "#/components/schemas/UserAdminDTO"
					}
				]
			},
			"TagPublicDTO": {
				"properties": {
					"addedAt": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"color",
					"name",
					"id"
				],
				"type": "object"
			},
			"MediaPublicDTO": {
				"properties": {
					"commentCount": {
						"type": "number",
						"format": "double"
					},
					"myRating": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"ratingCount": {
						"type": "number",
						"format": "double"
					},
					"ratingAvg": {
						"type": "number",
						"format": "double"
					},
					"favorite": {
						"type": "boolean"
					},
					"tags": {
						"items": {
							"$ref": "#/components/schemas/TagPublicDTO"
						},
						"type": "array"
					},
					"previewUrl": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string"
					},
					"isExplicit": {
						"type": "boolean"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"duration": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"height": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"width": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"size": {
						"type": "number",
						"format": "double"
					},
					"contentType": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"IMAGE",
							"VIDEO"
						]
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"commentCount",
					"myRating",
					"ratingCount",
					"ratingAvg",
					"favorite",
					"tags",
					"previewUrl",
					"createdAt",
					"isExplicit",
					"description",
					"duration",
					"height",
					"width",
					"size",
					"contentType",
					"type",
					"id"
				],
				"type": "object"
			},
			"MediaUserDTO": {
				"allOf": [
					{
						"$ref": "#/components/schemas/MediaPublicDTO"
					},
					{
						"properties": {
							"originalUrl": {
								"type": "string",
								"nullable": true
							}
						},
						"required": [
							"originalUrl"
						],
						"type": "object"
					}
				]
			},
			"MediaModeratorDTO": {
				"allOf": [
					{
						"$ref": "#/components/schemas/MediaUserDTO"
					},
					{
						"properties": {
							"previewKey": {
								"type": "string",
								"nullable": true
							},
							"originalKey": {
								"type": "string"
							},
							"deletedBy": {
								"type": "string",
								"nullable": true
							},
							"deletedAt": {
								"type": "string",
								"nullable": true
							},
							"moderationNotes": {
								"type": "string",
								"nullable": true
							},
							"moderatedById": {
								"type": "string",
								"nullable": true
							},
							"moderatedAt": {
								"type": "string",
								"nullable": true
							},
							"moderationStatus": {
								"type": "string",
								"enum": [
									"PENDING",
									"APPROVED",
									"REJECTED"
								]
							},
							"hash": {
								"type": "string"
							}
						},
						"required": [
							"previewKey",
							"originalKey",
							"deletedBy",
							"deletedAt",
							"moderationNotes",
							"moderatedById",
							"moderatedAt",
							"moderationStatus",
							"hash"
						],
						"type": "object"
					}
				],
				"description": "Sensitive/staff fields.\nShow them only if viewer can actually act as staff for media.\n\nIMPORTANT: \"read explicit\" is NOT a moderation entitlement by itself,\nso it must NOT unlock mod fields."
			},
			"MediaVisibleDTO": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/MediaPublicDTO"
					},
					{
						"$ref": "#/components/schemas/MediaUserDTO"
					},
					{
						"$ref": "#/components/schemas/MediaModeratorDTO"
					}
				]
			},
			"UserMediaListResponseDTO": {
				"properties": {
					"nextCursor": {
						"type": "string",
						"nullable": true
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/MediaVisibleDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"nextCursor",
					"data"
				],
				"type": "object"
			},
			"_36_Enums.CommentDeletedKind": {
				"type": "string",
				"enum": [
					"USER",
					"MODERATOR"
				]
			},
			"CommentDeletedKind": {
				"$ref": "#/components/schemas/_36_Enums.CommentDeletedKind"
			},
			"CommentAuthorDTO": {
				"properties": {
					"username": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"id"
				],
				"type": "object"
			},
			"CommentDTO": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/CommentAuthorDTO"
					},
					"createdAt": {
						"type": "string"
					},
					"deletedReason": {
						"type": "string",
						"nullable": true
					},
					"deletedKind": {
						"allOf": [
							{
								"$ref": "#/components/schemas/CommentDeletedKind"
							}
						],
						"nullable": true
					},
					"isDeleted": {
						"type": "boolean"
					},
					"content": {
						"type": "string",
						"nullable": true
					},
					"parentId": {
						"type": "string",
						"nullable": true
					},
					"userId": {
						"type": "string"
					},
					"mediaId": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"user",
					"createdAt",
					"deletedReason",
					"deletedKind",
					"isDeleted",
					"content",
					"parentId",
					"userId",
					"mediaId",
					"id"
				],
				"type": "object"
			},
			"UserCommentsListResponseDTO": {
				"properties": {
					"nextCursor": {
						"type": "string",
						"nullable": true
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/CommentDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"nextCursor",
					"data"
				],
				"type": "object"
			},
			"UserRatingItemDTO": {
				"properties": {
					"media": {
						"$ref": "#/components/schemas/MediaVisibleDTO"
					},
					"value": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"media",
					"value"
				],
				"type": "object"
			},
			"UserRatingsListResponseDTO": {
				"properties": {
					"nextCursor": {
						"type": "string",
						"nullable": true
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/UserRatingItemDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"nextCursor",
					"data"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"cookieAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "accessToken"
			},
			"optionalCookieAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "accessToken"
			}
		}
	},
	"info": {
		"title": "neobooru-backend",
		"version": "0.0.1",
		"contact": {}
	},
	"paths": {
		"/users/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserSelfDTO"
								}
							}
						}
					}
				},
				"description": "GET /users/me",
				"tags": [
					"Users"
				],
				"security": [
					{
						"cookieAuth": [
							"auth.load_permissions"
						]
					}
				],
				"parameters": []
			},
			"patch": {
				"operationId": "PatchMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserSelfDTO"
								}
							}
						}
					}
				},
				"description": "PATCH /users/me",
				"tags": [
					"Users"
				],
				"security": [
					{
						"cookieAuth": [
							"auth.load_permissions",
							"users.update_self"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUserPublic",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserVisibleDTO"
								}
							}
						}
					}
				},
				"description": "GET /users/:id",
				"tags": [
					"Users"
				],
				"security": [
					{
						"optionalCookieAuth": [
							"auth.load_permissions",
							"users.read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/{id}/uploads": {
			"get": {
				"operationId": "GetUserUploads",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMediaListResponseDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"optionalCookieAuth": [
							"auth.load_permissions",
							"users.read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"new",
								"old"
							]
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"IMAGE",
								"VIDEO"
							]
						}
					}
				]
			}
		},
		"/users/{id}/favorites": {
			"get": {
				"operationId": "GetUserFavorites",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMediaListResponseDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"optionalCookieAuth": [
							"auth.load_permissions",
							"users.read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"new",
								"old"
							]
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"IMAGE",
								"VIDEO"
							]
						}
					}
				]
			}
		},
		"/users/{id}/comments": {
			"get": {
				"operationId": "GetUserComments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserCommentsListResponseDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"optionalCookieAuth": [
							"auth.load_permissions",
							"users.read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"new",
								"old"
							]
						}
					}
				]
			}
		},
		"/users/{id}/ratings": {
			"get": {
				"operationId": "GetUserRatings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserRatingsListResponseDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"optionalCookieAuth": [
							"auth.load_permissions",
							"users.read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"new",
								"old"
							]
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"IMAGE",
								"VIDEO"
							]
						}
					}
				]
			}
		},
		"/users/me/avatar": {
			"post": {
				"operationId": "UploadMyAvatar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserSelfDTO"
								}
							}
						}
					}
				},
				"description": "POST /users/me/avatar",
				"tags": [
					"Users"
				],
				"security": [
					{
						"cookieAuth": [
							"auth.load_permissions",
							"users.avatar_update_self"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"avatar": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}